ApolloTardy
Completed Assignment:
This is the code for Pokemon Sword and Pokemon Shield. In this situation, you have entered a wild pokemon battle, and you are using a Pikachu. Condition one is when you attack with an electric-type move - for example, Thunderbolt - condition two is when your opponent attack you, condition three is if you try to run away, and condition four happens after condition one.